add sync

+++ aco.c
@@ -21,7 +21,7 @@
 // this header including should be at the last of the `include` directives list
 #include "aco_assert_override.h"
 
-int sysconf(int name){
+long sysconf(int name){
   if (name == _SC_PAGESIZE)
     return 4096;
   else return -1;
